Refactor isSanctionsMonitored to use uint256 for _days and update related calls for consistency across KintoID, IKintoID, and KintoIdHarness.

+++ test/harness/KintoIdHarness.sol
@@ -6,7 +6,7 @@ import {KintoID} from "@kinto-core/KintoID.sol";
 contract KintoIdHarness is KintoID {
     constructor(address _walletFactory, address _faucet) KintoID(_walletFactory, _faucet) {}
 
-    function isSanctionsMonitored(uint32) public view virtual override returns (bool) {
+    function isSanctionsMonitored(uint256) public view virtual override returns (bool) {
         return true;
     }
 }
